refactor some issues
