+++ components/datadog/actions/common/constants.mjs
@@ -1,9 +1,16 @@
+const MetricType = {
+  COUNT: "count",
+  GAUGE: "gauge",
+  RATE: "rate",
+};
+
 const metricTypes = [
-  "count",
-  "gauge",
-  "rate",
+  MetricType.COUNT,
+  MetricType.GAUGE,
+  MetricType.RATE,
 ];
 
 export default {
+  MetricType,
   metricTypes,
 };
